+++ redis/src/cluster_pipeline.rs
@@ -120,7 +120,7 @@ impl ClusterPipeline {
 
         from_redis_value(
             &(if self.commands.is_empty() {
-                Value::Bulk(vec![])
+                Value::Array(vec![])
             } else {
                 self.make_pipeline_results(con.execute_pipeline(self)?)
             }),