+++ redis/src/cluster_routing.rs
@@ -105,7 +105,7 @@ impl Routable for Cmd {
 impl Routable for Value {
     fn arg_idx(&self, idx: usize) -> Option<&[u8]> {
         match self {
-            Value::Bulk(args) => match args.get(idx) {
+            Value::Array(args) => match args.get(idx) {
                 Some(Value::Data(ref data)) => Some(&data[..]),
                 _ => None,
             },
@@ -115,7 +115,7 @@ impl Routable for Value {
 
     fn position(&self, candidate: &[u8]) -> Option<usize> {
         match self {
-            Value::Bulk(args) => args.iter().position(|a| match a {
+            Value::Array(args) => args.iter().position(|a| match a {
                 Value::Data(d) => d.eq_ignore_ascii_case(candidate),
                 _ => false,
             }),