+++ redis/tests/test_basic.rs
@@ -1258,7 +1258,7 @@ fn test_push_manager() {
         assert_eq!(
             (
                 PushKind::Invalidate,
-                vec![Value::Bulk(vec![Value::Data("key_1".as_bytes().to_vec())])]
+                vec![Value::Array(vec![Value::Data("key_1".as_bytes().to_vec())])]
             ),
             (kind, data)
         );
@@ -1272,7 +1272,7 @@ fn test_push_manager() {
     assert_eq!(
         (
             PushKind::Invalidate,
-            vec![Value::Bulk(vec![Value::Data("key_1".as_bytes().to_vec())])]
+            vec![Value::Array(vec![Value::Data("key_1".as_bytes().to_vec())])]
         ),
         (kind, data)
     );